新增訂單後台管理
order entity 增加付款狀態及出貨狀態
訂單管理可以查詢到訂單 但修改訂單不完整
購物車 action 暫時改接綠界 回傳post尚未完成

+++ workspace/springbootdemo/src/main/java/com/loveGod/demo/model/OrderDao.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface OrderDao extends JpaRepository<Order, Long> {
+public interface OrderDao extends JpaRepository<Order, Integer> {
 	
 	@Query(value="from Order  where userId = :userId")
 	public List <Order> findByUserId(String userId);
